ButonBotDjango/requirements.txt
Django==3.0.8
django-mptt==0.11.0
python-telegram-bot==12.8
redis
psycopg2-binary
psycopg2
gunicorn